+++ layer2/RepDihedral.c
@@ -319,7 +319,7 @@ Rep *RepDihedralNew(DistSet *ds)
 
         while(pos<length) {
 
-          mod_pos = fmod(pos + phase, dash_sum);
+          mod_pos = (float)fmod(pos + phase, dash_sum);
 
           VLACheck(I->V,float,(n*3)+5);
           
@@ -332,16 +332,16 @@ Rep *RepDihedralNew(DistSet *ds)
             cur_angle = angle * cons_pos1/length;
             
             v=I->V+n*3;
-            scale3f(x,cos(cur_angle),vx);
-            scale3f(y,sin(cur_angle),vy);
+            scale3f(x,(float)cos(cur_angle),vx);
+            scale3f(y,(float)sin(cur_angle),vy);
             add3f(vx,vy,v);
             add3f(a32,v,v);
             
             cur_angle = angle * cons_pos2/length;
             
             v+=3;
-            scale3f(x,cos(cur_angle),vx);
-            scale3f(y,sin(cur_angle),vy);
+            scale3f(x,(float)cos(cur_angle),vx);
+            scale3f(y,(float)sin(cur_angle),vy);
             add3f(vx,vy,v);
             add3f(a32,v,v);
